compile and stop scripts for productivity
I recommend putting these scripts in root so it's easier to access them so if you need to edit the project, you can access the root dir the easiest.

Note that the stop script returns you back to root once the dfs, master and slave have all been stopped.

If your project is not cloned into ~Documents/Github already, please do so. I have place resources in the Discord server about this modification.