[BUG FIX] Fixed formatting bug, logic ordering
Called reformatInput method after the while input is not a verified country loop to ensure the input is formatted correctly.
Previously, after entering an incorrect country/continent, the next input would be borked due to the placement of the reformatInput method call. Once placing it after the while loop to check for validity, the bug was squashed.